requirements.txt
flet[all]==0.27.6
pandas==2.2.3
numpy==2.2.2
pymongo==4.10.1
python-dotenv==1.0.1